fixed a bug that was preventing the build by jekyll
page-nav.html was missing a colon after the "append" calls (this may be an issue with how jupyter-book build works?), causing a generic jekyll error. The comment by 'chromatical' on https://talk.jekyllrb.com/t/liquid-error-wrong-number-of-arguments/732/6 suggested the answer - to just add those missing colons. Now make serve seems to work; and hopefully the build on github will as well.